Applying vertical responsivity in filter_screen
When I runned the application in my virtual device, I opened the Filter Screen, and it was with a error, because all the widgets weren't filled verticaly in my device.
So, I readjusted the distance between the widgets of the Column dinamicly with the dimension of the device on which the application will be running.